Modified useAuth.tsx to maintain userData, as well as some code refactoring
